feat: Remove AutomaticTranslationExtension portlet from bottom-all-container configuration - EXO-65108
