--- webapps/src/main/webapp/WEB-INF/conf/configuration.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd http://www.exoplatform.org/xml/ns/kernel_1_2.xsd"
-        xmlns="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd">
-
-    <import>war:/conf/automatic-translation/dynamic-container-configuration.xml</import>
-</configuration>
