--- webapps/src/main/webapp/html/automaticTranslationExtension.html
@@ -1,3 +1,0 @@
-<script type="text/javascript">
-  require(['PORTLET/automatic-translation/AutomaticTranslationExtension'], app => app.init());
-</script>
