+++ webapps/src/main/webapp/vue-apps/automatic-translation-extensions/activity-stream-extensions/components/ActivityCommentTranslatedBody.vue
@@ -1,5 +1,5 @@
 <!--
-Copyright (C) 2021 eXo Platform SAS.
+Copyright (C) 2024 eXo Platform SAS.
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU Affero General Public License as published by
